Print usage instead of panic for no arguments (#24)
* Print usage instead of panic for no arguments

Previously, if no arguments were passed to the
root command, a panic was triggered which in turn
bypassed the error handling strategy. Also the
provided message was not really helpful to the
user.

With this commit the usage is printed instead.
Additionally, an error is returned which is
correctly handled by the error handling strategy.

* Revert to old behavior

* Print usage if no subcommand was given

When no subcommand is passed to the application, only a short error text was printed to the console. Now the usage text is printed as well.

* Cleanup

* Add test for call without sub commands

* Remove explicit error discarding